backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
twilio==8.10.0
pymongo==4.6.1
Werkzeug==2.3.7
